\name{get_spectral}
\alias{get_spectral}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate Spectral Data
}
\description{
Generate a dataset with spectral distribution.
}
\usage{
get_spectral(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
Number of data to be generated
}
}
\value{
This function returns a dataframe with the spectral distribution
}
\references{
Bicici, E., & Yuret, D. (2007). Locally Scaled Density Based Clustering. International Conference on Adaptive and Natural Computing Algorithms (pp. 739-748). Berlin: Springer.
}
\author{
Fella Ulandari and Robert Kurniawan
}

\examples{
##Generate 1000 data##

get_spectral(1000)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{gendata}
\keyword{get_spectral}
